convert from int to float in the driver (working toward scaling factors)

+++ includes/fpga_bridge.hpp
@@ -48,7 +48,7 @@ public:
     int bulk_read_register(std::vector<uint32_t> address, std::vector<uint32_t> value);
     int start_capture(uint32_t n_buffers);
     int single_proxied_write_register(uint32_t proxy_address,uint32_t reg_address, uint32_t value);
-    int read_data(std::vector<uint32_t> &read_data);
+    int read_data(std::vector<float> &read_data);
     int apply_program(uint32_t address, std::vector<uint32_t> program);
     int set_channel_widths( std::vector<uint32_t> widths);
